requirements.txt
opencv-python
torch
torchvision
numpy
matplotlib
tqdm
yolov5
deepsort-realtime
